requirements-build.txt
pyinstaller>=6.20.0
pyarmor>=9.2.0,<9.3
pillow>=10.0.0
